DBService/requirements.txt
uvicorn
fastapi
redis
psycopg2-binary